controller unifies writing regular and masked values to system and user objects.
An object is a system or user object by virtue of where it is located (ancestor is system root or user root) and not specifically type of the object.  A method root_container() provides the corresponding root, for
any object, and this is used to determine if it is a system or user object.